requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
backoff>=2.2.1
tortoise-orm>=0.20.0
aiosqlite>=0.19.0
httpx>=0.26.0 